requirements.txt
# Python packages and versions used in development environment

alabaster==0.7.10
angles==2.0
appdirs==1.4.3
arrow==0.10.0
asn1crypto==0.22.0
attrs==18.2.0
Babel==2.5.1
basemap==1.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
black==18.6b4
bokeh==0.12.13
Bottleneck==1.2.1
cached-property==1.3.1
certifi==2018.8.13
cffi==1.11.2
chardet==3.0.4
circus==0.14.0
click==6.7
cliff==2.9.1
cloudpickle==0.5.2
cmd2==0.7.9
cmocean==1.2
coverage==4.4.2
cryptography==2.1.4
cycler==0.10.0
dask==0.16.1
defusedxml==0.5.0
distributed==1.20.2
docutils==0.14
driftwood==1.1.21
f90nml==1.0.2
feedgen==0.6.1
-e FVCOM_Cmd==0.1
gsw==3.2.0
h5netcdf==0.5.0
h5py==2.7.1
heapdict==1.0.0
idna==2.6
imagesize==0.7.1
iowait==0.2
isodate==0.6.0
Jinja2==2.10
locket==0.2.0
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==1.5.3
-e moad_tools==1.0.dev0
more-itertools==4.1.0
msgpack-python==0.5.1
-e NEMO-Cmd==1.3.dev0
-e NEMO-Nowcast==1.4
netCDF4==1.3.1
numpy==1.13.3
olefile==0.44
-e OPPTools==0.1.0
pandas==0.22.0
paramiko==2.3.1
partd==0.3.8
pbr==3.1.1
Pillow==4.3.0
pluggy==0.6.0
prettytable==0.7.2
psutil==5.4.0
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pygrib==2.0.2
PyNaCl==1.1.2
pyOpenSSL==17.4.0
pyparsing==2.2.0
pyperclip==1.6.0
pyproj==1.9.5.1
pyshp==1.2.12
PySocks==1.6.8
pytest==3.2.5
python-dateutil==2.6.1
python-hglib==2.4
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.2
raven==6.5.0
requests==2.18.4
requests-toolbelt==0.8.0
retrying==1.3.3
-e SalishSeaCmd==3.4.dev0
-e SalishSeaNowcast==3.3.dev0
-e SalishSeaTools==2.0
schedule==0.4.2
scipy==1.1.0
scour==0.36
Shapely==1.6.4.post1
six==1.11.0
snowballstemmer==1.2.1
sortedcontainers==1.5.7
Sphinx==1.7.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
stevedore==1.28.0
tblib==1.3.2
toml==0.9.6
toolz==0.8.2
tornado==4.5.3
typing==3.6.2
urllib3==1.22
utm==0.4.2
xarray==0.10.2
yapf==0.20.1
zeep==2.5.0
zict==0.1.3
